rename: components 경로 변경 및 주석 제거
